+++ main.c
@@ -121,7 +121,7 @@ init_compute_shader_ctx(ComputeShaderCtx *ctx, Arena a, uv3 rf_data_dim)
 {
 	ctx->rf_data_dim  = rf_data_dim;
 	/* TODO: send i16 data and convert to i32 on GPU */
-	size rf_raw_size     = rf_data_dim.w * rf_data_dim.h * rf_data_dim.d * sizeof(i32);
+	size rf_raw_size     = rf_data_dim.w * rf_data_dim.h * rf_data_dim.d * sizeof(i16);
 	size rf_decoded_size = rf_data_dim.w * rf_data_dim.h * rf_data_dim.d * sizeof(i32);
 
 	glGenBuffers(ARRAY_COUNT(ctx->rf_data_ssbos), ctx->rf_data_ssbos);
